requirements.txt
Flask==3.0.0
Werkzeug==2.2.2
Flask-Cors==3.0.10
google-cloud-storage==2.12.0
google-cloud-bigquery==3.12.0
pandas==1.4.4
firebase-admin==6.2.0
db-dtypes
itsdangerous==2.0.1
gunicorn